% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek-info.R
\name{current_gw}
\alias{current_gw}
\title{Get the Current Fantasy Premier League Gameweek}
\usage{
current_gw()
}
\value{
The current gameweek.
}
\description{
This function retrieves the current gameweek in the Fantasy Premier League.
}
\examples{
current_gw()

}
