% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek-info.R
\name{fpl_current_gw}
\alias{fpl_current_gw}
\title{Get the Current Fantasy Premier League Gameweek}
\usage{
fpl_current_gw()
}
\value{
The current FPL gameweek in character format.
}
\description{
This function retrieves the current gameweek in the Fantasy Premier League.
}
\examples{
fpl_current_gw()

}
