% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamteam-info.R
\name{fpl_dreamteam}
\alias{fpl_dreamteam}
\title{Get the Fantasy Premier League dream team for a specific gameweek.}
\usage{
fpl_dreamteam(gameweek_id = NULL)
}
\arguments{
\item{gameweek_id}{The ID of the gameweek for which to retrieve the dream team. If not provided,
the dream team for the current gameweek will be retrieved.}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{player}}{The names of the players in the dream team.}
\item{\code{points}}{The points scored by the player.}
\item{\code{team}}{The players' team.}
\item{\code{position}}{The players' position.}
}
}
\description{
This function retrieves the dream team (best performing players) for a specific
gameweek in the Fantasy Premier League using the official API. You can specify the
gameweek ID to get the dream team for a specific gameweek. If the gameweek ID is not
provided, the function will retrieve the dream team for the current gameweek.
}
\examples{
# Get the dream team for the current gameweek
fpl_dreamteam()

# Get the dream team for a specific gameweek (e.g., gameweek 5)
# fpl_dreamteam(gameweek_id = 5)

}
