% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-info.R
\name{fpl_find_player_id}
\alias{fpl_find_player_id}
\title{Find Fantasy Premier League player IDs.}
\usage{
fpl_find_player_id(name, team = NULL)
}
\arguments{
\item{name}{The name of the player to search for.}

\item{team}{(Optional) The team or teams to filter the search by. Defaults to all teams.}
}
\value{
The player IDs matching the search criteria. If no match is found, returns NA_integer_.
}
\description{
This function searches for player IDs in the Fantasy Premier League based on their names and optionally their teams. A fuzzy match is implemnted.
}
\examples{
fpl_find_player_id("Salah", team = "LIV")
fpl_find_player_id("Rashford")

}
