% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures-info.R
\name{fpl_fixtures_info}
\alias{fpl_fixtures_info}
\title{Get Fantasy Premier League fixtures information.}
\usage{
fpl_fixtures_info()
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{gw_id}}{The id for the gameweek.}
\item{\code{game_id}}{The fixture's season id.}
\item{\code{team_h}}{The home team.}
\item{\code{team_h_score}}{The home team's score.}
\item{\code{team_a}}{The away team.}
\item{\code{team_a_score}}{A numeric of the away team's score.}
\item{\code{stats}}{The fixture's statistics.}
\item{\code{kickoff_time}}{A datetime of the fixture's kickoff in UCT.}
\item{\code{kickoff_time_provisional}}{A logical indicating if the kickoff time is provisional.}
\item{\code{started}}{A logical indicating if the fixture has started.}
\item{\code{finished}}{A logical indicating if the fixture has finished.}
\item{\code{fpl_game_code}}{The fixture's FPL code.}
}
}
\description{
This function retrieves information about FPL fixtures, such as gameweek and fixture IDs, the teams, fixture score using the Fantasy Premier League API.
}
\examples{
fpl_team_info()

}
