% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek-info.R
\name{fpl_gameweek_info}
\alias{fpl_gameweek_info}
\title{Get Fantasy Premier League gameweek information.}
\usage{
fpl_gameweek_info()
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{gameweek}}{The name of the gameweek.}
\item{\code{gw_deadline}}{The deadline datetime of the gameweek in UTC.}
\item{\code{gw_finished}}{A logical value indicating whether the gameweek is finished.}
\item{\code{data_checked}}{A logical value indicating whether the data for the gameweek was checked.}
\item{\code{gw_current}}{A logical value indicating whether the gameweek is the active one.}
\item{\code{gw_previous}}{A logical value indicating whether the gameweek is the previous one.}
\item{\code{gw_next}}{A logical value indicating whether the gameweek is the next one.}
\item{\code{gw_highest_score}}{The highest score achieved by any FPL player in the gameweek.}
\item{\code{gw_average_score}}{The average score of all FPL entries in the gameweek.}
\item{\code{transfers_made}}{The total number of transfers made in the gameweek.}
\item{\code{chip_plays}}{The chip plays made in the gameweek.}
\item{\code{player_most_selected}}{The player who is most selected by FPL managers in the gameweek.}
\item{\code{player_most_captained}}{The player who is most captained by FPL managers in the gameweek.}
\item{\code{player_most_vice_captained}}{The player who is most vice-captained by FPL managers in the gameweek.}
\item{\code{player_most_transferred_in}}{The player who is most transferred-in by FPL managers in the gameweek.}
\item{\code{top_player}}{The top-performing player in the gameweek.}
\item{\code{top_player_info}}{Additional details about the top-performing player of that gameweek.}
}
}
\description{
This function retrieves information about the gameweeks and its details,
such as the gameweek name, deadline time, finished status, current status, previous status,
next status, highest score, average entry score, transfers made, chip plays, most selected player,
most captained player, most vice-captained player, most transferred-in player, and the top player
of the gameweek, along with their details.
}
\examples{
fpl_gameweek_info()

}
