% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-info.R
\name{fpl_league_info}
\alias{fpl_league_info}
\title{Get Fantasy Premier League classic league information.}
\usage{
fpl_league_info(league_id)
}
\arguments{
\item{league_id}{The ID of the classic league (numeric).}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{fpl_entry_id}}{The FPL entry ID of the participant.}
\item{\code{entry_season_id}}{The ID of the season for the participant.}
\item{\code{entry_name}}{The name of the participant.}
\item{\code{entry_team}}{The team name of the participant.}
\item{\code{gw_total}}{The total points scored by the participant in the current gameweek.}
\item{\code{overall_total}}{The overall total points scored by the participant.}
\item{\code{current_rank}}{The current rank of the participant in the league.}
\item{\code{previous_rank}}{The previous rank of the participant in the league.}
}
}
\description{
This function retrieves information about a classic league in Fantasy Premier League
using the provided league ID. It displays the league standings, including each
entry's total points, rank, and other relevant details.
}
\examples{
fpl_league_info(12345)

}
