% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager-info.R
\name{fpl_manager_history}
\alias{fpl_manager_history}
\title{Get Fantasy Premier League manager's historical performance of previous seasons.}
\usage{
fpl_manager_history(manager_id)
}
\arguments{
\item{manager_id}{The ID of the FPL manager (numeric).}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{season_name}}{The previous season the manager took part.}
\item{\code{total_points}}{The total points scored in that season.}
\item{\code{rank}}{The overall rank of that season.}
}
}
\description{
This function retrieves the historical performance an FPL manager (if exists).
}
\examples{
fpl_manager_history(123456)

}
