% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager-info.R
\name{fpl_manager_info}
\alias{fpl_manager_info}
\title{Get Fantasy Premier League manager information.}
\usage{
fpl_manager_info(manager_id)
}
\arguments{
\item{manager_id}{The ID of the FPL manager (nmeric).}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{gameweek_id}}{The ID of the gameweek.}
\item{\code{gw_points}}{The points scored by the manager in the gameweek.}
\item{\code{points_on_bench}}{The points scored by the manager's benched players in the gameweek.}
\item{\code{total_points}}{The total points scored by the manager.}
\item{\code{gw_rank}}{The rank of the manager in the gameweek.}
\item{\code{overall_rank}}{The overall rank of the manager.}
\item{\code{value}}{The value of the manager's team.}
\item{\code{bank}}{The remaining budget of the manager.}
\item{\code{gw_transfers}}{The number of transfers made by the manager in the gameweek.}
\item{\code{gw_transfers_cost}}{The total cost of transfers made by the manager in the gameweek.}
}
}
\description{
This function retrieves information about a Fantasy Premier League manager using
the provided manager ID. It displays the manager's performance details, including
points, rank, transfers, and other relevant statistics.
}
\examples{
fpl_manager_info(123456)

}
