% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek-info.R
\name{fpl_next_gw}
\alias{fpl_next_gw}
\title{Get the Next Fantasy Premier League Gameweek}
\usage{
fpl_next_gw()
}
\value{
The next gameweek in character format.
}
\description{
This function retrieves the next gameweek in the Fantasy Premier League.
}
\examples{
fpl_next_gw()

}
