% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-info.R
\name{fpl_player_fixture_stats}
\alias{fpl_player_fixture_stats}
\title{Get Fantasy Premier League player's fixture-level statistics.}
\usage{
fpl_player_fixture_stats(player_id)
}
\arguments{
\item{player_id}{The unique ID of the player in the FPL.}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{\code{gameweek}}{The gameweek number.}
\item{\code{game_id}}{The unique ID of the fixture.}
\item{\code{opponent}}{The opponent team's name.}
\item{\code{total_points}}{Total FPL points earned in the gameweek.}
\item{\code{starts}}{The number of starts in the gameweek.}
\item{\code{minutes}}{The total number of minutes played in the gameweek.}
\item{\code{goals_scored}}{The number of goals scored in the gameweek.}
\item{\code{assists}}{The number of assists provided in the gameweek.}
\item{\code{clean_sheets}}{The number of clean sheets in the gameweek.}
\item{\code{goals_conceded}}{The number of goals conceded in the gameweek.}
\item{\code{own_goals}}{The number of own goals scored in the gameweek.}
\item{\code{penalties_saved}}{The number of penalties saved in the gameweek.}
\item{\code{penalties_missed}}{The number of penalties missed in the gameweek.}
\item{\code{yellow_cards}}{The number of yellow cards received in the gameweek.}
\item{\code{red_cards}}{The number of red cards received in the gameweek.}
\item{\code{saves}}{The number of saves made in the gameweek.}
\item{\code{bonus}}{The bonus points earned in the gameweek.}
\item{\code{bps}}{The Bonus Points System (BPS) score in the gameweek.}
\item{\code{influence}}{The influence score in the gameweek.}
\item{\code{creativity}}{The creativity score in the gameweek.}
\item{\code{threat}}{The threat score in the gameweek.}
\item{\code{ict_index}}{The ICT (Influence, Creativity, Threat) index score in the gameweek.}
\item{\code{xg_scored}}{The expected goals scored in the gameweek.}
\item{\code{xa}}{The expected assists in the gameweek.}
\item{\code{xg_involvement}}{The expected goal involvements in the gameweek.}
\item{\code{xg_conceded}}{The expected goals conceded in the gameweek.}
\item{\code{price}}{The player's price for the gameweek.}
\item{\code{selected}}{The number of FPL managers who have selected the player for their team.}
\item{\code{transfers_in}}{The number of transfers in for the player in the gameweek.}
\item{\code{transfers_out}}{The number of transfers out for the player in the gameweek.}
\item{\code{transfers_balance}}{The net transfers balance for the player in the gameweek.}
}
}
\description{
This function retrieves detailed fixture-level statistics for a specific player in the Fantasy Premier League (FPL). You can obtain per-game season statistics including total points, starts, minutes played, goals scored, assists, clean sheets, goals conceded, and more.
}
\examples{
fpl_player_fixture_stats(player_id = 1)

}
