% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-info.R
\name{fpl_player_info}
\alias{fpl_player_info}
\title{Get Fantasy Premier League Information of All Players.}
\usage{
fpl_player_info()
}
\value{
A tibble (\code{data.frame}) with the following columns:
\describe{
\item{\code{player_season_id}}{The ID of the player for the current season.}
\item{\code{player_name}}{The web name of the player.}
\item{\code{first_name}}{The first name of the player.}
\item{\code{second_name}}{The last name of the player.}
\item{\code{team}}{The abbreviation of the team the player belongs to.}
\item{\code{squad_number}}{The squad number of the player.}
\item{\code{player_position}}{The player's position in the FPL game.}
\item{\code{player_fpl_code}}{The FPL code of the player.}
\item{\code{special}}{}
}
}
\description{
This function retrieves information about all FPL players, such as player IDs,
names, positions, and team details, using the Fantasy Premier League API.
}
\examples{
fpl_player_info()

}
