% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-info.R
\name{fpl_player_stats}
\alias{fpl_player_stats}
\title{Get Fantasy Premier League Headline Statistics of All Players.}
\usage{
fpl_player_stats()
}
\value{
A tibble containing player statistics with the following columns:
\describe{
\item{\code{player_season_id}}{Player ID for the current season.}
\item{\code{player_name}}{Name of the player.}
\item{\code{selected_pct}}{Percentage of FPL managers who have selected the player.}
\item{\code{selected_rank}}{Rank of the player based on selected percentage.}
\item{\code{selected_rank_type}}{Type of rank for selected percentage.}
\item{\code{current_price}}{Current price of the player in FPL.}
\item{\code{now_cost_rank}}{Rank of the player based on current price.}
\item{\code{now_cost_rank_type}}{Type of rank for current price.}
\item{\code{...}}{(additional columns with various player statistics).}
}
}
\description{
This function retrieves all players' headline statistics from the Fantasy Premier League API and
processes the data into a tibble with relevant player information and statistics.
}
\examples{
fpl_player_stats()

}
