% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager-info.R
\name{fpl_total_managers}
\alias{fpl_total_managers}
\title{Get the total number of Fantasy Premier League managers.}
\usage{
fpl_total_managers()
}
\value{
The total number of FPL managers.
}
\description{
This function retrieves the total number of managers participating in the Fantasy Premier League.
}
\examples{
fpl_total_managers()

}
