% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixtures.R
\name{get_fixtures}
\alias{get_fixtures}
\title{Get PL fixtures played and/or yet to play}
\usage{
get_fixtures()
}
\arguments{
\item{IDs}{a numeric with one or more team IDs}

\item{gameweeks}{range of gameweeks to retrieve picks for; if `NULL`, retrieves all gameweeks in season so far.}
}
\value{
a dataframe with two rows per fixture, one for the home team and one for the away team, including fixture difficulty (`difficulty`; scale of 1 - 5) and completed status (`finished`; TRUE or FALSE).
}
\description{
Get EPL fixtures played and/or yet to play in season, including fixture difficulty as estimated by the official FPL site.
}
\details{
Team ID can be found when examining points on the official FPL site, appearing after the `/entry/` part of the URL, e.g. `1` in `https://fantasy.premierleague.com/entry/1/event/12`
}
\examples{
# get completed fixtures
get_fixtures() \%>\%
  filter(finished == TRUE)

}
