% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures-info.R
\name{get_team_fixtures}
\alias{get_team_fixtures}
\title{Get Fixtures for Specific Premier League Teams}
\usage{
get_team_fixtures(team)
}
\arguments{
\item{team}{A character vector of team abbreviations.}
}
\value{
A tibble (data frame) with fixture information including game details.
}
\description{
This function retrieves fixture information for a specific Premier League teams.
}
\examples{
get_team_fixtures(c("ARS", "CHE"))

}
