% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures-info.R
\name{get_team_results}
\alias{get_team_results}
\title{Get Results for Specific Premier League Teams}
\usage{
get_team_results(team)
}
\arguments{
\item{team}{A character vector of team abbreviations.}
}
\value{
A tibble (data frame) with result information including game details.
}
\description{
This function retrieves result information for a specific Premier League teams.
}
\examples{
get_team_results(c("ARS", "CHE"))

}
