% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_picks.R
\name{manager_picks}
\alias{manager_picks}
\title{Get players picked each gameweek by each FPL manager}
\usage{
manager_picks(IDs, gameweeks = NULL)
}
\arguments{
\item{IDs}{a numeric with one or more team IDs}

\item{gameweeks}{range of gameweeks to retrieve picks for; if `NULL`, retrieves all gameweeks in season so far.}
}
\value{
a dataframe with each manager's team's information and total points
}
\description{
Get total points scored across the season so far by one or more FPL managers
}
\details{
Team ID can be found when examining points on the official FPL
site, appearing after the `/entry/` part of the URL, e.g. `1` in 
`https://fantasy.premierleague.com/entry/1/event/12`
}
\examples{
# first two gameweeks for two manager IDs
manager_picks(c(1, 2), 1:2)

}
