% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_scores.R
\name{manager_points}
\alias{manager_points}
\title{Get total points for each FPL manager}
\usage{
manager_points(IDs)
}
\arguments{
\item{IDs}{a numeric with one or more team IDs}
}
\value{
a dataframe with each manager's team's information and total points
}
\description{
Get total points scored across the season so far by one or more FPL managers
}
\details{
Team ID can be found when examining points on the official FPL
site, appearing after the `/entry/` part of the URL, e.g. `1` in
`https://fantasy.premierleague.com/entry/1/event/12`
}
\examples{
manager_points(c(1, 2))

}
