% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweek-info.R
\name{next_gw_deadline}
\alias{next_gw_deadline}
\title{Get the Deadline Time of the Next Fantasy Premier League Gameweek}
\usage{
next_gw_deadline()
}
\value{
The deadline time of the next gameweek in POSIXct format.
}
\description{
This function retrieves the deadline time of the next gameweek in the Fantasy Premier League.
}
\examples{
next_gw_deadline()

}
