% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_by_gameweek.R
\name{player_by_gw}
\alias{player_by_gw}
\title{Get points scored each gameweek for each FPL player}
\usage{
player_by_gw()
}
\value{
a dataframe with {n of gameweeks in season so far} rows per player containining all summary statistics related to player performance
}
\description{
Get gameweek-by-gameweek statistics for all FPL players using the official FPL API.
}
\examples{
player_by_gw()

}
