% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_summary.R
\name{player_points}
\alias{player_points}
\title{Get total points for each player}
\usage{
player_points()
}
\value{
a dataframe with one row per player containining all summary statistics related to player performance
}
\description{
Get summary statistics of all players, totalled across all gameweeks in season so far using the official FPL API.
}
\examples{
player_points()

}
