% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-colours.R
\docType{data}
\name{team_colours}
\alias{team_colours}
\title{Colours of the Premier League 23/24}
\format{
\subsection{\code{team_colours}}{

A data frame of the 23/24 Premier League teams:
\describe{
\item{team_abb}{Team's abbriviation}
\item{primary_colour}{Team's primary colour}
\item{secondary_colour}{Team's secondary colour}
}
}
}
\source{
Data are collected from 20 Premier League Teams style guides.
}
\usage{
team_colours
}
\description{
Colours of the Premier League 23/24
}
\examples{
team_colours

}
\keyword{datasets}
