% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpts_h2h.R
\name{xpts_h2h}
\alias{xpts_h2h}
\title{Calculate expected points for a head-to-head league table.}
\usage{
xpts_h2h(leagueid)
}
\arguments{
\item{leagueid}{an integer with a single valid head-to-head league ID}
}
\value{
a dataframe with expected points (xPts) and expected rank (xRank)
added to actual points, actual rank, and team information
}
\description{
Calculate expected points (xPts) and expected rank (xRank) for
each team in a head-to-head table as the average points (i.e. 0, 1, 3)
expected against all teams in that league each gameweek (including AVERAGE 
team, present if odd number of teams in the league), not just the actual
opponent. Requires entering username and password in interactive dialogue
in order to access private leagues. Uses the [`fplscrapR`](https://github.com/wiscostret/fplscrapR).
}
\details{
League ID can be found when examining the head-to-head table on the
official FPL site, appearing after the `/leagues/` part of the URL, e.g. `1`
in `https://fantasy.premierleague.com/leagues/1/standings/h`
}
\examples{
xpts_h2h(1)

}
\seealso{
`fplscrapR` package by [wiscostret](https://github.com/wiscostret/fplscrapR)
}
