% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl-table.R
\name{pl_table}
\alias{pl_table}
\title{Print Premier League Table.}
\usage{
pl_table()
}
\value{
A \code{gt} table object displaying the Premier League table.
}
\description{
This function prints a table with Premier League teams' performance data, including
the number of matches played, wins, draws, losses and points.
}
\examples{
pl_table()

}
